feat: add cmd/tracee
Add binary with subcommands for collecting (ebpf), and
rules.
