+++ builder/Dockerfile.alpine-tracee-container
@@ -101,6 +101,7 @@ COPY . /tracee
 
 RUN make clean && \
     BTFHUB=$BTFHUB make tracee-ebpf && \
+    BTFHUB=$BTFHUB make tracee && \
     make tracee-rules && \
     make rules && \
     rm -rf ./3rdparty/btfhub/ && \
@@ -118,9 +119,10 @@ WORKDIR /tracee
 
 COPY --from=tracee-make /tracee/dist/tracee-ebpf /tracee
 COPY --from=tracee-make /tracee/dist/tracee-rules /tracee
+COPY --from=tracee-make /tracee/dist/tracee /tracee/tracee
 COPY --from=tracee-make /tracee/dist/rules/ /tracee/rules/
 COPY --from=tracee-make /tracee/builder/entrypoint.sh /tracee/entrypoint.sh
-COPY --from=tracee-make /tracee/cmd/tracee-rules/templates/ /tracee/templates/
+COPY --from=tracee-make /tracee/cmd/tracee/rules/templates/ /tracee/templates/
 
 ENTRYPOINT ["/tracee/entrypoint.sh"]
 
@@ -137,9 +139,10 @@ WORKDIR /tracee
 COPY --from=tracee-make /tracee /tracee/src
 COPY --from=tracee-make /tracee/dist/tracee-ebpf /tracee
 COPY --from=tracee-make /tracee/dist/tracee-rules /tracee
+COPY --from=tracee-make /tracee/dist/tracee /tracee
 COPY --from=tracee-make /tracee/dist/rules/ /tracee/rules
 COPY --from=tracee-make /tracee/builder/entrypoint.sh /tracee/entrypoint.sh
-COPY --from=tracee-make /tracee/cmd/tracee-rules/templates/ /tracee/templates/
+COPY --from=tracee-make /tracee/cmd/tracee/rules/templates/ /tracee/templates/
 
 ENTRYPOINT ["/tracee/entrypoint.sh"]
