--- cmd/tracee-rules/testdata/signatures/helpers.rego
@@ -1,1 +1,0 @@
-../../../../signatures/rego/helpers.rego
