+++ cmd/tracee/collect/capabilities_test.go
@@ -1,4 +1,4 @@
-package main
+package collect
 
 import (
 	"testing"
