+++ cmd/tracee/collect/flags/errors.go
