+++ cmd/tracee/collect/initialize/kernelconfig.go
