+++ cmd/tracee/collect/internal/printer/printer.go
