+++ cmd/tracee/rules/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package rules
 
 import (
 	"bytes"
