+++ cmd/tracee/rules/input.go
@@ -1,4 +1,4 @@
-package main
+package rules
 
 import (
 	"bufio"
