+++ cmd/tracee/rules/signature.go
@@ -1,4 +1,4 @@
-package main
+package rules
 
 import (
 	_ "embed"
