+++ cmd/tracee/rules/templates/csv.tmpl
