+++ cmd/tracee/rules/templates/falcosidekick.tmpl
