+++ cmd/tracee/rules/templates/rawjson.tmpl
