+++ cmd/tracee/rules/templates/simple.tmpl
