+++ cmd/tracee/rules/templates/sprig.tmpl
