+++ cmd/tracee/rules/testdata/goldens/broken.tmpl
