+++ cmd/tracee/rules/testdata/signatures/anti_debugging_ptraceme.rego
@@ -1,0 +1,1 @@
+../../../../../signatures/rego/anti_debugging_ptraceme.rego
