+++ cmd/tracee/rules/testdata/signatures/code_injection.rego
@@ -1,0 +1,1 @@
+../../../../../signatures/rego/code_injection.rego
