+++ cmd/tracee/rules/testdata/signatures/helpers.rego
@@ -1,0 +1,1 @@
+../../../../../signatures/rego/helpers.rego
